feat(api): add deletion case for certification related badge

+++ api/lib/domain/usecases/delete-unassociated-badge.js
@@ -1,10 +1,15 @@
-const { BadgeForbiddenDeletionError } = require('../errors');
+const { AcquiredBadgeForbiddenDeletionError, CertificationBadgeForbiddenDeletionError } = require('../errors');
 
 module.exports = async function deleteUnassociatedBadge({ badgeId, badgeRepository }) {
   const isAssociated = await badgeRepository.isAssociated(badgeId);
+  const isRelatedToCertification = await badgeRepository.isRelatedToCertification(badgeId);
 
   if (isAssociated) {
-    throw new BadgeForbiddenDeletionError();
+    throw new AcquiredBadgeForbiddenDeletionError();
+  }
+
+  if (isRelatedToCertification) {
+    throw new CertificationBadgeForbiddenDeletionError();
   }
 
   return badgeRepository.delete(badgeId);
